Delete GlotPress meta data on uninstall
Fixes #71.

+++ uninstall.php
@@ -8,6 +8,7 @@
 namespace Required\Traduttore;
 
 use WP_Filesystem_Base;
+use wpdb;
 
 defined( 'WP_UNINSTALL_PLUGIN' ) || exit;
 
@@ -17,16 +18,27 @@ global $wp_filesystem;
 if ( ! $wp_filesystem ) {
 	require_once ABSPATH . '/wp-admin/includes/admin.php';
 
-	if ( ! \WP_Filesystem() ) {
-		return false;
-	}
+	\WP_Filesystem();
 }
 
-array_map(
-	function ( $file_or_folder ) use ( $wp_filesystem ) {
-		$wp_filesystem->delete( $file_or_folder, true, is_dir( $file_or_folder ) ? 'd' : 'f' );
-	},
-	glob( get_temp_dir() . 'traduttore-*' )
-);
+if ( $wp_filesystem ) {
+	array_map(
+		function ( $file_or_folder ) use ( $wp_filesystem ) {
+			$wp_filesystem->delete( $file_or_folder, true, is_dir( $file_or_folder ) ? 'd' : 'f' );
+		},
+		glob( get_temp_dir() . 'traduttore-*' )
+	);
 
-$wp_filesystem->rmdir( ZipProvider::get_cache_dir(), true );
+	$wp_filesystem->rmdir( ZipProvider::get_cache_dir(), true );
+}
+
+/* @var wpdb $wpdb */
+global $wpdb;
+
+$meta_key_prefix = '_traduttore_';
+
+// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
+$query = $wpdb->prepare( "DELETE FROM `$wpdb->gp_meta` WHERE `meta_key` LIKE %s ", $wpdb->esc_like( $meta_key_prefix ) . '%' );
+
+// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
+$wpdb->query( $query );
